Module owlc_parser::ast::statements

Structs

Block
Represents a block. e.g. {...}. A Block creates a new lexical scope. Variables declared inside the Block are not visible outside the Block.

CompilationUnit
Represents a collection of functions that are included in the compilation unit.

FnProto
Represents a function prototype (not the function body). e.g. fn f(x).

Stmt
Represents a statement. Statements do not have any value unlike Expr. Statements usually have side effects.

Enums

StmtKind
Internal representation for Stmt.