Enum owlc_parser::ast::statements::StmtKind

pub enum StmtKind {
    Block(Block),
    Fn {
        proto: FnProto,
        body: Option<Block>,
    },
    While {
        condition: Expr,
        body: Block,
    },
    For,
    IfElse {
        if_condition: Expr,
        if_body: Block,
        else_body: Option<Block>,
    },
    Let {
        iden: String,
        initializer: Expr,
    },
    Return {
        value: Expr,
    },
    ExprSemi {
        expr: Expr,
    },
    Noop,
}

Internal representation for Stmt.

Variants

Block(Block)

Wrapper around Block.

Fn

Represents a function declaration (with definition if not extern). e.g. fn f(x) { return 3; } Field body.kind should always be variant StmtKind::Block. If field body is None, function is an extern function.

Grammar

fn = ( fn proto, block ) | (fn proto extern, ";" );

Fields of Fn

proto: FnProtobody: Option<Block>

Field should be None if function is an extern fn.

While

Fields of While

condition: Expr

while condition.

body: Block

Content of the while loop.

For

IfElse

Represents an if and else statement. The else block is optional.

Grammar

if = "if", expression, block, [ else ];
else = "else", block;

Fields of IfElse

if_condition: Exprif_body: Blockelse_body: Option<Block>

Else block is option.

Let

Fields of Let

iden: Stringinitializer: Expr

Return

Fields of Return

value: Expr

ExprSemi

Represents an expression with semi colon (expression with side effect).

Fields of ExprSemi

expr: Expr

Noop

No-op (used for empty REPL input).