Trait owlc_parser::visitor::AstVisitor

pub trait AstVisitor {
    fn visit_expr(&mut self, expr: &Expr) { ... }

    fn visit_literal_expr(&mut self, _val: &i64) { ... }

    fn visit_identifier_expr(&mut self, _iden: &str) { ... }

    fn visit_func_call(&mut self, _callee: &str, _args: &[Expr]) { ... }

    fn visit_bin_expr(&mut self, lhs: &Expr, rhs: &Expr, _op_type: &TokenKind) { ... }

    fn visit_stmt(&mut self, stmt: &Stmt) { ... }

    fn visit_compilation_unit(&mut self, compilation_unit: &CompilationUnit) { ... }

    fn visit_fn_proto(&mut self, _ident: &str, _args: &[String]) { ... }

    fn visit_block(&mut self, block: &Block) { ... }

    fn visit_fn_stmt(&mut self, proto: &FnProto, body: &Option<Block>) { ... }

    fn visit_while_stmt(&mut self, condition: &Expr, body: &Block) { ... }

    fn visit_for_stmt(&mut self) { ... }

    fn visit_if_else_stmt(
        &mut self, 
        if_condition: &Expr, 
        if_body: &Block, 
        else_body: &Option<Block>
    ) { ... }

    fn visit_let_stmt(&mut self, _ident: &str, initializer: &Expr) { ... }

    fn visit_return_stmt(&mut self, val: &Expr) { ... }

    fn visit_expr_semi_stmt(&mut self, expr: &Expr) { ... }
}

Provided methods

fn visit_expr(&mut self, expr: &Expr)

Visits any Expr node. Uses pattern matching to dispatch the correct visit method. This method should not be overridden unless custom behavior is needed.

fn visit_literal_expr(&mut self, _val: &i64)

fn visit_identifier_expr(&mut self, _iden: &str)

fn visit_func_call(&mut self, _callee: &str, _args: &[Expr])

fn visit_bin_expr(&mut self, lhs: &Expr, rhs: &Expr, _op_type: &TokenKind)

fn visit_stmt(&mut self, stmt: &Stmt)

Visits any Stmt node. Uses pattern matching to dispatch the correct visit method. This method should not be overridden unless custom behavior is needed.

fn visit_compilation_unit(&mut self, compilation_unit: &CompilationUnit)

fn visit_fn_proto(&mut self, _ident: &str, _args: &[String])

fn visit_block(&mut self, block: &Block)

fn visit_fn_stmt(&mut self, proto: &FnProto, body: &Option<Block>)

Panics

This method should panic if param body.kind is not a StmtKind::Block variant.

fn visit_while_stmt(&mut self, condition: &Expr, body: &Block)

fn visit_for_stmt(&mut self)

fn visit_if_else_stmt(
&mut self,
if_condition: &Expr,
if_body: &Block,
else_body: &Option<Block>
)

fn visit_let_stmt(&mut self, _ident: &str, initializer: &Expr)

fn visit_return_stmt(&mut self, val: &Expr)

fn visit_expr_semi_stmt(&mut self, expr: &Expr)

Loading content...